requirements.txt
redfish
httplib2
python_hpilo
configparser
